+++ naga/src/front/spv/mod.rs
@@ -1633,6 +1633,7 @@ impl<I: Iterator<Item = u32>> Frontend<I> {
                 | S::SubgroupBallot { .. }
                 | S::SubgroupCollectiveOperation { .. }
                 | S::SubgroupGather { .. }
+                | S::SubgroupBallotFindBit { .. }
                 | S::RayPipelineFunction(..) => {}
                 S::Call {
                     function: ref mut callee,